pyproject.toml
[tool.poetry]
name = "cronvisio"
version = "0.9.0"
description = "cronvisio - Monitor and control services through cron and notifiers."
authors = ["Albert Weichselbraun <albert@weichselbraun.net>"]
license = "GPL-3.0"
readme = "README.md"
homepage = "http://github.com/AlbertWeichselbraun/cronitor"
repository = "http://github.com/AlbertWeichselbraun/cronitor"
keywords = ["bibPublish", "bibtex", "publish"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "^3.8 || ^3.9 || ^3.10 || ^3.11 || ^3.12"
matrix-nio = "^0.24.0"

[tool.poetry.scripts]
cronvisio = "cronvisio.cli:cli"

# code formatting with black
[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
extend-exclude = '\.html$|\.json$|\.txt$|\.gz$'
include = '''
  ^/src/|^/tests/|^/benchmarking/|^/examples/
'''
